backend/requirements.txt

Flask
Flask-Cors
tensorflow
numpy
pandas
Pillow
Werkzeug
scikit-learn
joblib
langchain-google-genai
google-generativeai
